LayoutTests/fast/page-color-sampling/color-sampling-full-viewport-backdrop-filter-expected.txt

Content underneath backdrop filter

PASS colors.top is "multiple"
PASS colors.left is null
PASS colors.right is null
PASS colors.bottom is "multiple"
PASS successfullyParsed is true

TEST COMPLETE
